Server/requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
pymysql
cryptography
bcrypt
pydantic
pydantic[email]
pyyaml
python-jose[cryptography]
